added authorization for routes

+++ src/main/java/com/postInfo/sys/post/controller/admin/AdminController.java
@@ -1,6 +1,7 @@
 package com.postInfo.sys.post.controller.admin;
 
 import com.postInfo.sys.post.data.UserData;
+import com.postInfo.sys.post.data.response.MessageResponse;
 import com.postInfo.sys.post.data.response.SuccessResponse;
 import com.postInfo.sys.post.model.ERole;
 import com.postInfo.sys.post.model.Role;
@@ -31,8 +32,10 @@ public class AdminController {
 
     @GetMapping("")
     @PreAuthorize("hasRole('ADMIN')")
-    public List<UserData> findAll() {
+    public ResponseEntity findAll() {
         List<User> user = userService.findAll();
+        if(user == null)
+            return ResponseEntity.ok(new MessageResponse("No users found"));
         List<UserData> users = new ArrayList<>();
         for(final User singleUser: user) {
             UserData userData = new UserData();
@@ -42,23 +45,26 @@ public class AdminController {
             userData.setRole(singleUser.getRoles());
             users.add(userData);
         }
-        return users;
+        return ResponseEntity.ok(users);
     }
 
     @RequestMapping(value = "/role/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public List<UserData> findUserByRole(@PathVariable String id) {
+    public ResponseEntity findUserByRole(@PathVariable String id) {
         Role role = null;
         if(id.compareTo("user") == 0)
             role = roleService.findByName(ERole.ROLE_USER);
         else if (id.compareTo("admin") == 0)
             role = roleService.findByName(ERole.ROLE_ADMIN);
         else
-            throw new RuntimeException("Error: Role is not found");
+            return ResponseEntity.badRequest().body(new MessageResponse("Role is not found"));
 
         List<User> user = userService.findUserByRoles(role);
-        List<UserData> users = new ArrayList<>();
 
+        if(user == null)
+            return ResponseEntity.badRequest().body(new SuccessResponse(id,"No user exists for given role"));
+
+        List<UserData> users = new ArrayList<>();
         for(final User singleUser: user) {
             UserData userData = new UserData();
             userData.setId(singleUser.getId());
@@ -67,25 +73,29 @@ public class AdminController {
             userData.setRole(singleUser.getRoles());
             users.add(userData);
         }
-        return users;
+        return ResponseEntity.ok(users);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     @PreAuthorize("hasRole('ADMIN')")
-    public UserData getUserById(@PathVariable("id") String id) {
+    public ResponseEntity getUserById(@PathVariable("id") String id) {
         User user = userService.findUserById(id);
+        if(user == null)
+            return ResponseEntity.badRequest().body(new SuccessResponse(id,"No user found with given id"));
         UserData userData = new UserData();
         userData.setId(user.getId());
         userData.setUsername(user.getUsername());
         userData.setEmail(user.getEmail());
         userData.setRole(user.getRoles());
-        return userData;
+        return ResponseEntity.ok(userData);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<?> deleteUser(@PathVariable String id) {
         User user = userService.findUserById(id);
+        if(user == null)
+            return ResponseEntity.badRequest().body(new SuccessResponse(id, "No user found with given id"));
         userService.delete(user);
         return ResponseEntity
                 .ok(new SuccessResponse(id, user.getUsername() + " deleted Successfully"));
